[qa] cause test_framework to mine the initial chain *now*
Summary:
Currently, the test mines blocks back in Jan 1 2014 to
generate an initial blockchain.  Then tests zip into the future
to mine blocks in the current time.  This change fixes it, so that
the initial blockchain is generated recently, but allows it to be
overriden for specific tests.

Test Plan:
  ./test/functional/test_runner.py

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1424